Started working on Bonfire: Exact Change
Bonfire: Exact Change is an Advanced Algorithm Bonfire exercise from
Free Code Camp.
http://www.freecodecamp.com/challenges/bonfire-exact-change

The aim of this exercise is to design a cash register drawer function
that accepts purchase price, payment, and cash-in-drawer.

Cash-in-drawer, or cid, is a 2d array listing available currency.

The function returns three different values for giving change:
1. "Insufficient Funds", string, if there is not enough for change
2. "Closed", string, if the cid empties out
3. Array, an array of change of coins and bills sorted out from highest
   to lowest.