Reworked entire testing procedure.
The original procedure involves setting a variable with
recognized keywords representing the different suites.
This limits the testing to suites and does not allow the
testing and picking of individual files.

The new procedure makes use of an environment variable set by
    export TESTS="keyword:1,2,3;keyword2:all;keyword3:4...6"

This env variable TESTS is a strin,  parsed by the test
master file to select the tests to be processed.

Suites are set by recognized keywords:
basic
intermediate
upper
advanced
all

followed by a ":" and a comma delimited list of file numbers.
A range may also specified by using an ellipsis.

Example:
    export TESTS="basic:5...10"

Will set the test collection to files 5, 6, 7, 8, 9, 10 from the
basic suite.

To specify multiple suites, the keywords are separated by
a semicolon.

Example:
    export TESTS="upper:1,2,3;intermediate:all"

This will set the test collection to files 1, 2, 3 for the upper
suite and all of the intermediate suite.

"all" is a special keyword that can be specified alone or with
another keyword.

Example:
    export TESTS="all" // will test all suites
	 export TESTS="basic:all" // will test all basic suite tests