Bonfire: Friendly Date Ranges
This program converts two dates into a "friendly" date
range that is presented to the user.

This program is broken up into many different functions,
each (hopefully) focused on a single task.

makeOrdinal(day)
1. day, number, the date to be converted into ordinal
   number (1st, 2nd, 3rd, 4th, ...)

monthName(month)
1. month, number, the 0 based index of the month to be
   converted to its own name, i.e. 0 = January, 1 = February

makeDate(dates)
1. dates, array, a two element array with the values for
   starting date and end date

Rather than initialise the variables in the main friendly()
function, it is done here instead to keep the statement
count to a minimum and the functions focused on their own
task.

same(type, dates)
1. type, string, a comma delimited string of the types to
   test for. Valid types are 'month, date, year'
2. dates, array, a two element array with the values for
   starting date and end date

Rather than a verbose IF statement comparing months, dates,
and years, this function returns a boolean instead after
the comparison of the specified types.

make(type, dates)
1. type, string, a comma delimited string of the types to
   make from the specified date.
2. dates, array, a two element array containing the values
   for the start and end date

This function is responsible for formatting the returned values
needed. It assembles it from the type specified.

yearOrMore(dates)
1. dates, array, a two element array containing the values
   for the start and the end date

If the dates can't be handled by same(), because of calculations
needed, yearOrMore() is then called. They are similar otherwise.

friendly(dates)
1. dates, array, a two element array containing the values
   for the start and the end date.

friendly() is the "main" entry function of the whole collection.
It accepts an array of two date ranges in the format of
'YYYY-MM-DD' and then returns them in a 'user friendly' format.